EnemyHealth with Prefab on Death
This file is to be used for Little Melee System asset. To use add to Armored Stand in demo scene. Create a new DamageInfoBody and DamageInfoArms script. Just use the original DamageInfo script for this. Copy, Paste and rename public class to correct name. Remove old DamgeInfo on body parts and add new scripts. Choose your Dead GameObject. Now the enemy should die when health gets to 0 with new object in its place.